Fix argument order of embed.
In LaTeX, optional arguments are enclosed in square brackets, and always
come _before_ required arguments. That's why e.g. a cube root is written
as
```
\sqrt[3]{1+x}
```
and not
```
\sqrt{1+x}[3]
```
which is actually also valid latex that displays a square root followed
by a 3 wrapped in square brackets.

The embed feature got this backwards. Fix it to use the standard order.